services/opcua-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
asyncua==1.1.5
cryptography
python-jose[cryptography]
prometheus-client
influxdb-client[async]
psycopg2-binary
